+++ lib/src/progressDialog.dart
@@ -94,6 +94,9 @@ class ProgressDialog implements _ProgressDialog {
   final Duration? transitionDuration;
 
   bool _show = false;
+
+  bool get isShowed => _show;
+
   _ProgressDialogWidget? _progressDialogWidget;
 
   ProgressDialog(this.context,
@@ -105,8 +108,8 @@ class ProgressDialog implements _ProgressDialog {
       this.onCancel,
       this.dismissable,
       this.onDismiss,
-      this.title,
-      this.message,
+      @required this.title,
+      @required this.message,
       this.dialogStyle,
       this.transitionDuration}) {
     _initProgress();
@@ -174,7 +177,7 @@ class ProgressDialog implements _ProgressDialog {
 
   ///future function let you show ProgressDialog until future (param)
   ///reach the end of its action
-  static Future future<T>(
+  static Future<T?> future<T>(
     BuildContext context, {
     @required Future? future,
     DialogStyle? dialogStyle,
@@ -185,8 +188,8 @@ class ProgressDialog implements _ProgressDialog {
     OnProgressCancel? onProgressCancel,
     Function? onDismiss,
     bool? dismissable,
-    Widget? message,
-    Widget? title,
+    @required Widget? message,
+    @required Widget? title,
     Widget? cancelText,
     Widget? progressWidget,
     DialogTransitionType? dialogTransitionType,
@@ -214,13 +217,11 @@ class ProgressDialog implements _ProgressDialog {
 
     var output;
     await future.then((data) {
-      if (onProgressFinish != null)
-        onProgressFinish = onProgressFinish?.call(data);
+      if (onProgressFinish != null) onProgressFinish = onProgressFinish?.call(data);
       output = data;
       pDialog.dismiss();
     }).catchError((error) {
-      if (onProgressError != null)
-        onProgressError = onProgressError?.call(error);
+      if (onProgressError != null) onProgressError = onProgressError?.call(error);
       pDialog.dismiss();
     });
 
@@ -271,8 +272,7 @@ class _ProgressDialogWidget extends StatefulWidget {
   }
 }
 
-class _ProgressDialogWidgetState extends State<_ProgressDialogWidget>
-    implements _ProgressDialog {
+class _ProgressDialogWidgetState extends State<_ProgressDialogWidget> implements _ProgressDialog {
   Widget? _title, _message, _loading;
   Color? _backgroundColor;
 
@@ -283,8 +283,7 @@ class _ProgressDialogWidgetState extends State<_ProgressDialogWidget>
 
     Widget? title = _title ?? widget.title;
     Widget message = _message ?? (widget.message ?? SizedBox.shrink());
-    Color backgroundColor =
-        _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
+    Color backgroundColor = _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
     Widget loading = (_loading ?? widget.loadingWidget) ??
         Container(
           padding: EdgeInsets.all(10.0),
@@ -331,9 +330,7 @@ class _ProgressDialogWidgetState extends State<_ProgressDialogWidget>
               Expanded(
                 child: DefaultTextStyle(
                   child: Semantics(child: message),
-                  style: (widget.dialogStyle?.contentTextStyle ??
-                          dialogTheme.contentTextStyle) ??
-                      (theme.textTheme.subtitle1 ?? TextStyle()),
+                  style: (widget.dialogStyle?.contentTextStyle ?? dialogTheme.contentTextStyle) ?? (theme.textTheme.subtitle1 ?? TextStyle()),
                 ),
               ),
             ],
@@ -412,6 +409,7 @@ class CustomProgressDialog implements _CustomProgressDialog {
   final Duration? transitionDuration;
 
   bool _show = false;
+  bool get isShowed => _show;
   _CustomProgressDialogWidget? _progressDialogWidget;
 
   CustomProgressDialog(
@@ -509,12 +507,10 @@ class CustomProgressDialog implements _CustomProgressDialog {
     var output;
     try {
       await future.then((data) {
-        if (onProgressFinish != null)
-          onProgressFinish = onProgressFinish?.call(data);
+        if (onProgressFinish != null) onProgressFinish = onProgressFinish?.call(data);
         output = data;
       }).catchError((error) {
-        if (onProgressError != null)
-          onProgressError = onProgressError?.call(error);
+        if (onProgressError != null) onProgressError = onProgressError?.call(error);
       });
     } catch (e) {}
     pDialog.dismiss();
@@ -531,8 +527,7 @@ class _CustomProgressDialogWidget extends StatefulWidget {
   final double? blur;
   final Color? backgroundColor;
   final bool? dismissable;
-  _CustomProgressDialogWidgetState _dialogWidgetState =
-      _CustomProgressDialogWidgetState();
+  _CustomProgressDialogWidgetState _dialogWidgetState = _CustomProgressDialogWidgetState();
 
   _CustomProgressDialogWidget({
     Key? key,
@@ -560,16 +555,13 @@ class _CustomProgressDialogWidget extends StatefulWidget {
   }
 }
 
-class _CustomProgressDialogWidgetState
-    extends State<_CustomProgressDialogWidget>
-    implements _CustomProgressDialog {
+class _CustomProgressDialogWidgetState extends State<_CustomProgressDialogWidget> implements _CustomProgressDialog {
   Widget? _loadingWidget;
   Color? _backgroundColor;
 
   @override
   Widget build(BuildContext context) {
-    Color backgroundColor =
-        _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
+    Color backgroundColor = _backgroundColor ?? (widget.backgroundColor ?? generalBarrierColor);
     Widget loadingWidget = (this._loadingWidget ?? widget.loadingWidget) ??
         Container(
           padding: EdgeInsets.all(10.0),
@@ -589,8 +581,7 @@ class _CustomProgressDialogWidgetState
       onDismiss: widget.onDismiss,
       barrierColor: backgroundColor,
       dialog: Padding(
-        padding: MediaQuery.of(context).viewInsets +
-            const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
+        padding: MediaQuery.of(context).viewInsets + const EdgeInsets.symmetric(horizontal: 20.0, vertical: 24.0),
         child: Center(
           child: loadingWidget,
         ),